Fix hr and h? tag (without br in tail)
- Pretty interface of try_parse_md_atx_heading
-  A few tests does not  passed, which should be fixed  as soon as possiable.

+++ include/pltxt2htm/astnode/node_type.hh
@@ -8,7 +8,7 @@ namespace pltxt2htm {
  * @brief ast node type of Quantum-Physics (aka. Physics-Lab, pl) text
  * @details show details in astnode.hh, every AstNode owns a NodeType
  */
-enum class NodeType : ::std::uint_least32_t {
+enum class NodeType : ::std::size_t {
     base = 0,
     // a utf-8 char
     u8char,
